requirements.txt
python-dotenv==1.0.1
ragas==0.2.9
langchain==0.3.6
nltk==3.9.1